+++ CHANGELOG.rst
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog`_,
 and this project adheres to `Semantic Versioning`_.
 
+Unreleased
+----------
+
+* Added: to_array conversion to Buffer class
+
 
 Version 0.2.3 (2019-11-22)
 --------------------------